% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcHazardDetected}
\alias{CalcHazardDetected}
\title{Compute hazard of each detection within time-step}
\usage{
CalcHazardDetected(data, dt, transdat, parameter, observer_speed, type,
  hzfn)
}
\arguments{
\item{data}{(x, y, t) data matrix}

\item{dt}{time step}

\item{transdat}{transect data matrix}

\item{parameter}{(scale, shape, diffusion) parameters}

\item{observer_speed}{speed of observer}

\item{type}{1 = point, 0 = line transect}

\item{hzfn}{hazard function code (see ?hazardfns)}
}
\value{
PDF for within-timestep detection
}
\description{
Compute hazard of each detection within time-step
}
