% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcInitialDistribution}
\alias{CalcInitialDistribution}
\title{Calculates the initial distribution of animal locations.
Assumes uniform distribution relative to transect.}
\usage{
CalcInitialDistribution(num_cells, delta, region_size)
}
\arguments{
\item{num_cells}{vector with number of cells in (total space, x-direction,
y-direction)}

\item{delta}{spatial and temporal increments (dx, dt)}

\item{region_size}{size of survey region in (x,y) extents}
}
\value{
Row vector with i^th entry probability animal in i^th grid cell initially
}
\description{
Calculates the initial distribution of animal locations.
Assumes uniform distribution relative to transect.
}
