% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcSurvivalPr}
\alias{CalcSurvivalPr}
\title{Computes the probability of survival for each spatial location}
\usage{
CalcSurvivalPr(t, parameter, num_cells, delta, strip_size, buffer,
  observer_speed, type, hzfn, nint = 1L)
}
\arguments{
\item{t}{time step}

\item{parameter}{(scale, shape, diffusion) parameter}

\item{num_cells}{number of cells in (x, y, all) dimensions}

\item{delta}{(dx, dt) vector}

\item{strip_size}{size of strip in (x, y) dimensions}

\item{buffer}{buffer size}

\item{observer_speed}{speed of the observer}

\item{type}{transect type}

\item{hzfn}{hazard function code}

\item{nint}{not used

 @return row vector of survival probabilities over space}
}
\description{
Computes the probability of survival for each spatial location
}
