% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcTrm}
\alias{CalcTrm}
\title{Calculates the sparse transition rate matrix}
\usage{
CalcTrm(num_cells, sd, dx)
}
\arguments{
\item{num_cells}{vector with number of cells in (total space, x-direction,
y-direction)}

\item{sd}{vector of diffusive standard deviation for each behavioural state}

\item{dx}{grid cell size in the space (c.f. delta(0))}
}
\value{
sparse transition rate matrix
}
\description{
Calculates the sparse transition rate matrix
}
