% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Detect}
\alias{Detect}
\title{Thins probability distribution by the proportion detected in each grid cell}
\usage{
Detect(t, pr, parameter, num_cells, delta, strip_size, buffer,
  observer_speed, type, hzfn)
}
\arguments{
\item{t}{time step}

\item{pr}{probability distribution over finite grid}

\item{parameter}{(scale, shape, diffusion) parameter}

\item{num_cells}{number of cells in (x, y, all) dimensions}

\item{delta}{(dx, dt) vector}

\item{strip_size}{size of strip in (x, y) dimensions}

\item{buffer}{buffer width}

\item{observer_speed}{speed of the observer}

\item{type}{transect type}

\item{hzfn}{hazard function code 

@return  thinned probability distribution}
}
\description{
Thins probability distribution by the proportion detected in each grid cell
}
