% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Diffuse}
\alias{Diffuse}
\title{Diffuse probability distribution over space}
\usage{
Diffuse(a, v, t, num_cells, krylov_dim = 30L, tol = 1e-10)
}
\arguments{
\item{a}{transition rate matrix}

\item{v}{vector to be multiplied}

\item{t}{time to diffuse over}

\item{num_cells}{vector with number of cells in (total space, x-direction,
y-direction)}

\item{krylov_dim}{dimension of the approximating Krylov space}

\item{tol}{tolerance in error}
}
\value{
diffused probability distribution
}
\description{
Calculate product of v with matrix exponential of a using
the Arnoldi process. Thereby diffusing the probability distribution
according to Brownian motion. Code is transcribed from Expokit package.
}
\note{
NOTICE
Permission to use, copy, modify, and distribute EXPOKIT and its
  supporting documentation for non-commercial purposes, is hereby
    granted without fee, provided that this permission message and
    copyright notice appear in all copies. Approval must be sought for
      commercial purposes as testimony of its usage in applications.
    
    Neither the Institution (University of Queensland) nor the Author
      make any representations about the suitability of this software for
        any purpose.  This software is provided ``as is'' without express or
       implied warranty.
      
      The work resulting from EXPOKIT has been published in ACM-Transactions 
        on Mathematical Software, 24(1):130-156, 1998.
      
      The bibtex record of the citation:
        
        ARTICLE{EXPOKIT,
                 AUTHOR  = {Sidje, R. B.},
                 TITLE   = {{Expokit.} {A} Software Package for
                   Computing Matrix Exponentials},
                   JOURNAL = {ACM Trans. Math. Softw.},
                   VOLUME  = {24},
                   NUMBER  = {1},
                   PAGES   = {130-156}
          YEAR    = {1998}
        }
      
      Certain elements of the current software may include inadequacies
        that may be corrected at any time, as they are discovered. The Web 
        always contains the latest updates.
      
      Original Author:
        Roger B. Sidje <rbs@maths.uq.edu.au>
        Department of Mathematics, University of Queensland 
        Brisbane, QLD-4072, Australia, (c) 1996-2006 All Rights Reserved
}
