% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise.R
\name{Discretise}
\alias{Discretise}
\title{Discretise transects and map detections to spatial grid}
\usage{
Discretise(ds)
}
\arguments{
\item{ds}{distance sampling data (see mds function)}
}
\value{
list of two matrices:
  $dtrans matrix with (transect ID, number of observations made, total time steps)
  $ddata contains row for each detection with (transect id, grid cell
    detection was recorded in, x distance, y distance, exact time of detection)
  $numcells contains number of grid cells in total space, x-distance, y-distance
  $ymax maximum distance in the y direction (usually longest line transect or radial distance for points)
  $T length of survey in time steps
}
\description{
Discretise transects and map detections to spatial grid
}
