% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unload.R
\name{.onUnload}
\alias{.onUnload}
\title{Unload moveds package compiled code}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{lib path}
}
\value{
none
}
\description{
Unload moveds package compiled code
}
