% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{EstDiff}
\alias{EstDiff}
\title{Estimate diffusion rate from tag data}
\usage{
EstDiff(tags)
}
\arguments{
\item{tags}{list of matrices, one for each tagged animal, with (x, y, t) rows}
}
\value{
estimated diffusion rate
}
\description{
Estimate diffusion rate from tag data
}
