% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mds_line}
\alias{example_mds_line}
\title{Example line transect data with movement}
\format{A list of two data frames 
\describe{
  \item{obs}{sightings with transect, x, y, t records} 
  \item{trans}{entry for each transect surveyed: ID, effort}
}}
\usage{
example_mds_line
}
\description{
A dataset containing observations from a simulated line transect survey 
where animals move by Brownian motion.
}
\keyword{datasets}
