% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mds_point_mods}
\alias{example_mds_point_mods}
\title{Example model fits to example_mds_point data}
\format{An object of class \code{list} of length 3.}
\usage{
example_mds_point_mods
}
\description{
A list of cds1d, cds2d, and mds2d models.
}
\keyword{datasets}
