% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetHist}
\alias{GetHist}
\title{Computes PDF of observed detections for each (x,y) cell around the observer.}
\usage{
GetHist(working_parameter, range, transdat, auxiliary_data, delta,
  num_cells, T, ymax, buffer, fixed_sd = 0, hzfn = 1L,
  move_method = 1L)
}
\arguments{
\item{working_parameter}{unconstrained version of parameter vector containing
(detection shape, detection scale, diffusion sd)}

\item{range}{to compute out to in x and y directions}

\item{transdat}{matrix with (stripsize(1), numcells in y, totaltimestep, number of observations)}

\item{auxiliary_data}{vector containing (area x extent, area y extent, strip width, transect_type)}

\item{delta}{vector of (dx, dt) spacetime increments}

\item{num_cells}{number of cells in (total space, x-direction, y-direction)}

\item{T}{total time of survey for longest transect}

\item{ymax}{maximum length of a transect}

\item{buffer}{buffer distance}

\item{fixed_sd}{if move_method = 2}

\item{hzfn}{hazard function code (see ?hazardfns)}

\item{move_method}{0 = 2d CDS model, 1 = 2d MDS model (movement estimated),
2 = 2d MDS model (movement fixed)}
}
\value{
matrix where (i,j) entry is cell i*dx perpendicular and j*dx forward of
  observer
unpack auxiliary data
}
\description{
Computes PDF of observed detections for each (x,y) cell around the observer.
}
