% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetObserverPosition}
\alias{GetObserverPosition}
\title{Get observation location in 2D space}
\usage{
GetObserverPosition(time, strip_size, buffer, delta, transect_type,
  observer_speed)
}
\arguments{
\item{time}{time to return observer position}

\item{strip_size}{size of strip in (x, y) dimensions}

\item{buffer}{buffer size}

\item{delta}{(dx, dt) vector}

\item{transect_type}{0 = line, 1 = point}

\item{observer_speed}{speed of observer}
}
\value{
(x, y) location of observer at time t
}
\description{
Get observation location in 2D space
}
