% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPenc}
\alias{GetPenc}
\title{Computes covered area for entire survey}
\usage{
GetPenc(working_parameter, transdat, auxiliary_data, delta, num_cells, T,
  ymax, buffer, fixed_sd, hzfn, move_method)
}
\arguments{
\item{working_parameter}{unconstrained version of parameter vector containing
(detection shape, detection scale, diffusion sd)}

\item{transdat}{matrix with (stripsize(1), numcells in y, totaltimestep, number of observations)}

\item{auxiliary_data}{vector containing (area x extent, area y extent, strip width, transect_type)}

\item{delta}{vector of (dx, dt) spacetime increments}

\item{num_cells}{number of cells in (total space, x-direction, y-direction)}

\item{T}{total time of survey for longest transect}

\item{ymax}{maximum length of a transect}

\item{buffer}{buffer distance}

\item{fixed_sd}{if move_method = 2}

\item{hzfn}{hazard function code (see ?hazardfns)}

\item{move_method}{0 = 2d CDS model, 1 = 2d MDS model (movement estimated),
2 = 2d MDS model (movement fixed)}
}
\value{
negative log-likelihood

covered area
unpack auxiliary data
}
\description{
Computes covered area for entire survey
}
