% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InTransect}
\alias{InTransect}
\title{Computes what grid cells are inside and outside transect}
\usage{
InTransect(num_cells, strip_size, dx, w, ymax, buffer, type)
}
\arguments{
\item{num_cells}{number of cells in (total, x, y) direction}

\item{strip_size}{size of strip in (x,y) directions}

\item{dx}{grid cell size}

\item{w}{for lines, half-width, for points radius}

\item{ymax}{maximum forward distance for lines}

\item{buffer}{distance}

\item{type}{=0 for lines, =1 for points}
}
\value{
vector with 1 for each grid cell inside and 0 otherwise
}
\description{
Computes what grid cells are inside and outside transect
}
