% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.mds}
\alias{logLik.mds}
\title{Return loglikelihood of a fitted MDS model}
\usage{
\method{logLik}{mds}(object, ...)
}
\arguments{
\item{object}{fitted MDS model}

\item{...}{ignored}
}
\value{
loglikelihood with attribute df = number of parameters
}
\description{
Return loglikelihood of a fitted MDS model
}
