% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{mds.gof}
\alias{mds.gof}
\title{Chi-squared Goodness-of-fit tests}
\usage{
mds.gof(mod, delta = NULL, yquant = 0.95)
}
\arguments{
\item{mod}{fitted mds model (returned from mds function)}

\item{delta}{use a different (dx, dt); otherwise uses mod$ds$delta}

\item{yquant}{quantile of forward distance to truncate at (for robustness to outliers)}
}
\value{
plots of empirical and estimated CDFs for distances and list of outputs from ks.test function
}
\description{
Chi-squared Goodness-of-fit tests
}
