% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mds}
\alias{mds}
\title{Fit distance sampling with movement model}
\usage{
mds(ds, move, start, print = FALSE, level = 0.95, ...)
}
\arguments{
\item{ds}{named list of distance sampling data:
$data: matrix with observations for each row with (transect ID, x, y, t) columns
$transect: matrix with transects for each row with (transect ID, length)
$aux: vector of variables (region width, region length, truncation distance, observer speed, transect type 0line 1point)
$delta: vector of discretisation sizes in (space, time)
$hazardfn: code for hazard function to use (see ?hazardfns), default is 1
$move: 0 = 2d CDS model, 1 = 2d MDS model with estimated diffusion (must supply tag data), 2 = 2D MDS model with fixed diffusion (must supply fixed.sd)}

\item{move}{named list of movement data:
$fixed.sd fixed diffusion paramter, must be supplied and only used when ds$move = 2
$data list with matrix for each individual tagged with (x, y, t) observations in each row, required for ds$move = 1}

\item{start}{named initial values for (detection scale, detection shape, diffusion rate (ds$move == 1))}

\item{print}{FALSE by default, if TRUE then likelihood and parameters are printed after each evaluation}

\item{level}{confidence interval level, default is 0.95}

\item{...}{additional arguments to be passed to nlm (the optimisation routine)}
}
\value{
Named list:
        $result table of estimated parameters, standard errors, and confidence intervals
        $cor estimated correlation matrix between parameter estimates
        $penc average probability of detection
        $AIC Akaike's information criterion score for the model
        $fit the fitted model object output by optim
        $ds ds argument 
        $move move argument 
        $level confidence level supplied
}
\description{
Fit distance sampling with movement model
}
