% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Natural2Working}
\alias{Natural2Working}
\title{Transform natural parameters to unconstrained working parameters}
\usage{
Natural2Working(parameter, hzfn = 1L)
}
\arguments{
\item{parameter}{natural parameters}

\item{hzfn}{hazard function type}
}
\value{
working parameters
}
\description{
Transform natural parameters to unconstrained working parameters
}
