% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NegativeLogLikelihood}
\alias{NegativeLogLikelihood}
\title{Computes negative log-likelihood of moveDs model}
\usage{
NegativeLogLikelihood(working_parameter, start, data, transdat,
  auxiliary_data, delta, num_cells, T, ymax, buffer, movement_data,
  fixed_sd = 0, hzfn = 1L, move_method = 1L, print = FALSE,
  con = 100)
}
\arguments{
\item{working_parameter}{unconstrained version of parameter vector containing
(detection shape, detection scale, diffusion sd)}

\item{start}{start value for parameters on natural scale}

\item{data}{matrix with (trans id, grid cell,t) distance sampling survey data (assumed to be ordered by transect and time)}

\item{transdat}{matrix with (stripsize(1), numcells in y, totaltimestep, number of observations)}

\item{auxiliary_data}{vector containing (area x extent, area y extent, strip width, transect_type)}

\item{delta}{vector of (dx, dt) spacetime increments}

\item{num_cells}{number of cells in (total space, x-direction, y-direction)}

\item{T}{total time of survey for longest transect}

\item{ymax}{maximum length of a transect}

\item{buffer}{buffer distance}

\item{movement_data}{field object where each component represents an individual
path and contains a matrix where each row is an observed location (x,y,t)}

\item{fixed_sd}{if move_method = 2}

\item{hzfn}{hazard function code (see ?hazardfns)}

\item{move_method}{0 = 2d CDS model, 1 = 2d MDS model (movement estimated),
2 = 2d MDS model (movement fixed)}

\item{print}{if TRUE then print likelihood and parmeters after evaluation}

\item{con}{parameters are constrained to be between 1/con * start value and 
con * start value}
}
\value{
negative log-likelihood
}
\description{
Computes negative log-likelihood of moveDs model
}
