% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mds}
\alias{plot.mds}
\title{Plots fitted PDF from MDS model}
\usage{
\method{plot}{mds}(x, delta = NULL, yquant = 0.95, nint = 1000, ...)
}
\arguments{
\item{x}{fitted mds model (returned from mds function)}

\item{delta}{use a different (dx, dt); otherwise uses mod$ds$delta}

\item{yquant}{maximum y distance quantile to plot up to}

\item{nint}{discrete spacing to plot approximating spline with}

\item{...}{ignored}
}
\description{
Plots fitted PDF from MDS model
}
