% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{point_movedat}
\alias{point_movedat}
\title{Example animal movement data}
\format{A data frames with columns (x, y, t)}
\usage{
point_movedat
}
\description{
A dataset containing observations from a simulated telemetry survey 
on ten tagged animals moving by Brownian motion.
}
\keyword{datasets}
