% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mds}
\alias{predict.mds}
\title{Fit distance sampling with movement model}
\usage{
\method{predict}{mds}(object, newdata = NULL, stratum.name = NULL,
  level = 0.95, ...)
}
\arguments{
\item{object}{fitted mds model}

\item{newdata}{if NULL then predicted abundance for all covered area, 
otherwise, provide transects (in form of ds$transect described 
for mds function) for transects covering stratum of interest}

\item{stratum.name}{optional, used to label prediction object}

\item{level}{confidence interval level, default is 0.95}

\item{...}{ignored}
}
\value{
Named list:
        $result table of estimated parameters, standard errors, and confidence intervals
        $cor estimated correlation matrix between parameter estimates
        $penc average probability of detection
        $AIC Akaike's information criterion score for the model
        $fit the fitted model object output by optim
        $ds ds argument 
        $move move argument 
        $level confidence level supplied
}
\description{
Fit distance sampling with movement model
}
