% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{s2sigmab}
\alias{s2sigmab}
\title{Convert moveds s,d parameters to Hayes-Buckland b, sigma}
\usage{
s2sigmab(s, d, v = 1, T = 1)
}
\arguments{
\item{s}{scale parameter}

\item{d}{shape parameter}

\item{v}{observer speed, default is one, if zero then point transect assumed}

\item{T}{duration of point transect}
}
\value{
named vector of shape (b) and scale (sigma) under Hayes-Buckland
}
\description{
Convert moveds s,d parameters to Hayes-Buckland b, sigma
}
