% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{SimulateMovementData}
\alias{SimulateMovementData}
\title{Simulate movement paths from Brownian motion}
\usage{
SimulateMovementData(num.animals, observation.times, diffusion)
}
\arguments{
\item{num.animals}{number of tagged animals to simulate}

\item{observation.times}{vector of times animal locations are recorded}

\item{diffusion}{diffusion rate}
}
\value{
list of matrices, each matrix corresponds to an individual and has
        columns (xlocation, ylocation, observation time)
}
\description{
Simulate movement paths from Brownian motion
}
