% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mds}
\alias{summary.mds}
\title{Summary of MDS fitted model}
\usage{
\method{summary}{mds}(object, ...)
}
\arguments{
\item{object}{MDS fitted model}

\item{...}{ignored}
}
\value{
print summary; N and D are abundance and density estimates. Reported
abundance is for the region of area given by ds$aux[1:2]
}
\description{
Summary of MDS fitted model
}
