% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Working2Natural}
\alias{Working2Natural}
\title{Transform working parameters (for the optimiser) to natural parameters}
\usage{
Working2Natural(working_parameter, hzfn = 1L)
}
\arguments{
\item{working_parameter}{working parameters}

\item{hzfn}{hazard function type}
}
\value{
natural parameters
}
\description{
Transform working parameters (for the optimiser) to natural parameters
}
